requirements.txt
astroid==2.14.2
attrs==22.2.0
black @ git+https://github.com/psf/black@d9b8a6407e2f46304a8d36b18e4a73d8e0613519
blinker==1.5
click==8.1.3
dill==0.3.6
Faker==17.0.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.16.2
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==2.0.3
tomlkit==0.11.6
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.14.1